src/ec2/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
flask
requests
opentelemetry-exporter-otlp
boto3
